ModelTraining/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
azure-cosmos
fastai==1.0.61
librosa==0.8.0
pydub==0.24.1
pandas
numpy
torchaudio==0.6.0
git+https://github.com/fastaudio/fastai_audio@0.1
ipython
spacy